+++ game/slot/megajack/aztecgold/aztecgold_calc.go
@@ -68,17 +68,17 @@ func CalcStat(ctx context.Context, mrtp float64) float64 {
 	var s slot.Stat
 
 	var calc = func(w io.Writer) float64 {
-		var reshuf = float64(s.Count())
+		var reshuf = s.Count()
 		var cost, _ = g.Cost()
 		var lrtp, srtp = s.LineRTP(cost), s.ScatRTP(cost)
 		var rtpsym = lrtp + srtp
-		var qmjap = float64(s.BonusCount(mjap)) / reshuf
+		var qmjap = s.BonusCount(mjap) / reshuf
 		var rtpmjap = Ebon * qmjap * 100
 		var rtp = rtpsym + rtpmjap
 		fmt.Fprintf(w, "symbols: %.5g(lined) + %.5g(scatter) = %.6f%%\n", lrtp, srtp, rtpsym)
-		fmt.Fprintf(w, "pyramid bonuses: frequency 1/%.5g, rtp = %.6f%%\n", reshuf/float64(s.BonusCount(mjap)), rtpmjap)
+		fmt.Fprintf(w, "pyramid bonuses: frequency 1/%.5g, rtp = %.6f%%\n", reshuf/s.BonusCount(mjap), rtpmjap)
 		if s.JackCount(mjj) > 0 {
-			fmt.Fprintf(w, "jackpots: count %d, frequency 1/%.12g\n", s.JackCount(mjj), reshuf/float64(s.JackCount(mjj)))
+			fmt.Fprintf(w, "jackpots: count %g, frequency 1/%.12g\n", s.JackCount(mjj), reshuf/s.JackCount(mjj))
 		}
 		fmt.Fprintf(w, "RTP = %.5g(sym) + %.5g(mjap) = %.6f%%\n", rtpsym, rtpmjap, rtp)
 		return rtp
