+++ game/slot/megajack/slotopol/slotopol_calc.go
@@ -91,20 +91,20 @@ func CalcStat(ctx context.Context, mrtp float64) float64 {
 	var s slot.Stat
 
 	var calc = func(w io.Writer) float64 {
-		var reshuf = float64(s.Count())
+		var reshuf = s.Count()
 		var cost, _ = g.Cost()
 		var lrtp, srtp = s.LineRTP(cost), s.ScatRTP(cost)
 		var rtpsym = lrtp + srtp
-		var qmje9 = float64(s.BonusCount(mje9)) / reshuf / float64(g.Sel)
+		var qmje9 = s.BonusCount(mje9) / reshuf / float64(g.Sel)
 		var rtpmje9 = Emje * 9 * qmje9 * 100
-		var qmjm = float64(s.BonusCount(mjm)) / reshuf / float64(g.Sel)
+		var qmjm = s.BonusCount(mjm) / reshuf / float64(g.Sel)
 		var rtpmjm = Emjm * qmjm * 100
 		var rtp = rtpsym + rtpmje9 + rtpmjm
 		fmt.Fprintf(w, "symbols: %.5g(lined) + %.5g(scatter) = %.6f%%\n", lrtp, srtp, rtpsym)
-		fmt.Fprintf(w, "spin9 bonuses: frequency 1/%.5g, rtp = %.6f%%\n", reshuf/float64(s.BonusCount(mje9)), rtpmje9)
-		fmt.Fprintf(w, "monopoly bonuses: frequency 1/%.5g, rtp = %.6f%%\n", reshuf/float64(s.BonusCount(mjm)), rtpmjm)
+		fmt.Fprintf(w, "spin9 bonuses: frequency 1/%.5g, rtp = %.6f%%\n", reshuf/s.BonusCount(mje9), rtpmje9)
+		fmt.Fprintf(w, "monopoly bonuses: frequency 1/%.5g, rtp = %.6f%%\n", reshuf/s.BonusCount(mjm), rtpmjm)
 		if s.JackCount(mjj) > 0 {
-			fmt.Fprintf(w, "jackpots: count %d, frequency 1/%.12g\n", s.JackCount(mjj), reshuf/float64(s.JackCount(mjj)))
+			fmt.Fprintf(w, "jackpots: count %g, frequency 1/%.12g\n", s.JackCount(mjj), reshuf/s.JackCount(mjj))
 		}
 		fmt.Fprintf(w, "RTP = %.5g(sym) + %.5g(mje9) + %.5g(mjm) = %.6f%%\n", rtpsym, rtpmje9, rtpmjm, rtp)
 		return rtp
