+++ game/slot/novomatic/powerstars/powerstars_calc.go
@@ -94,7 +94,7 @@ func CalcStatStars(ctx context.Context, wc2, wc3, wc4 bool) float64 {
 		go slot.Progress(ctx2, &s, calc)
 		BruteForceStars(ctx2, &s, g, reels, wc2, wc3, wc4)
 		var dur = time.Since(t0)
-		var comp = float64(s.Reshuf(1)) / float64(s.Planned()) * 100
+		var comp = s.Reshuf(1) / s.Planned() * 100
 		fmt.Printf("completed %.5g%%, selected %d lines, time spent %v\n", comp, g.GetSel(), dur)
 	}()
 	return calc(os.Stdout)
