+++ game/slot/playtech/greatblue/greatblue_calc.go
@@ -41,11 +41,11 @@ func CalcStat(ctx context.Context, mrtp float64) float64 {
 	var s slot.Stat
 
 	var calc = func(w io.Writer) float64 {
-		var reshuf = float64(s.Count())
+		var reshuf = s.Count()
 		var cost, _ = g.Cost()
 		var lrtp, srtp = s.LineRTP(cost), s.ScatRTP(cost)
 		var rtpsym = lrtp + srtp
-		var fghits = float64(s.FreeHits())
+		var fghits = s.FreeHits()
 		var fsavr1, multavr = FirstSreespins()
 		var q = fghits * fsavr1 / reshuf
 		var sq = 1 / (1 - fghits*multavr*15/reshuf)
